requirements.txt
selenium==4.15.2
pytest==7.4.3
allure-pytest==2.13.2
requests==2.31.0
pytest-html==4.1.1
python-dotenv~=1.1.0
allure-python-commons~=2.13.2
webdriver-manager~=4.0.2